Add 2399.cpp

+++ leetcode/2399.cpp
@@ -6,7 +6,7 @@ public:
     bool checkDistances(string s, vector<int>& distance) {
         vector<int> firstpos(26, -1);
 
-        for(int i = 0; i < s.size(); i++){
+        for(int i = 0; i < (int)s.size(); i++){
             int idx = s[i] - 'a';
             if(firstpos[idx] == -1){
                 firstpos[idx] = i;
